Created RestFul API'S and Added Exception Handling.

+++ airBnbApp/airBnbApp/src/main/java/com/example/Project_Rishit/airBnbApp/dto/RoomResponseDto.java
@@ -3,12 +3,14 @@ package com.example.Project_Rishit.airBnbApp.dto;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @RequiredArgsConstructor
 public class RoomResponseDto {
     private Long roomId;
     private String type;
-    private Double basePrice;
+    private BigDecimal basePrice;
     private String[] photos;
     private String[] amenities;
     private Integer capacity;
